Use CreatureAttackTacticalObjective. This is probably making ExpAi very bad ATM (Striking is basically counted twice)
git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4263 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/ai/CreatureAttackTacticalObjective.java
@@ -32,7 +32,7 @@ class CreatureAttackTacticalObjective extends AbstractTacticalObjective
     private final AbstractAI ai;
     private final BattleEvalConstants bec;
 
-    CreatureAttackTacticalObjective(int priority, Client client,
+    CreatureAttackTacticalObjective(float priority, Client client,
             Legion liveLegion, Creature creature, AbstractAI ai,
             BattleEvalConstants bec)
     {
